pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi[all]>=0.104.1",
    "google-generativeai>=0.8.4",
    "llama-index>=0.12.30",
    "llama-index-embeddings-gemini>=0.3.2",
    "llama-index-embeddings-google-genai>=0.1.0",
    "llama-index-llms-gemini>=0.4.14",
    "llama-index-llms-google-genai>=0.1.7",
    "llama-index-llms-groq>=0.3.1",
    "llama-index-readers-file>=0.4.7",
    "llama-index-vector-stores-chroma>=0.4.1",
    "llama-index-vector-stores-supabase>=0.3.0",
    "passlib[bcrypt]==1.7.4",
    "pdf2image>=1.17.0",
    "pydantic>=2.4.2",
    "pydantic-settings>=2.0.3",
    "pymupdf>=1.25.5",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "sqlmodel>=0.0.11",
    "uvicorn>=0.24.0",
]
